Allow using various git diff options for comparisons
Instead of having to do funky stuff with soft git resets, this allows the user
to pass a `--diff` option.  This option can be anything that works with both
`git diff-tree` and `git rev-list`.  This option does not work with mercurial
yet.  This is particularly helpful in CI pipelines that rely on PRs.
Effectively, you run `git lint --diff 'HEAD ^<target-branch>'`

The change basically works by looking at all the commits that are different
instead of just the `--last-commit`.  It uses the `git rev-list` command to
determine which commits are different.  It then uses the same `modified_lines`
function with a slight modification.  `modified lines` now uses a regex OR
condition for all of the commits instead of just one commit.